[refactor] Move KernelProfilerCUDA into backend/cuda, and make CUPTI toolkit as one of its member variables
