+++ taichi/backends/cuda/cuda_context.cpp
@@ -76,11 +76,8 @@ void CUDAContext::launch(void *func,
 
   KernelProfilerBase::TaskHandle task_handle;
   // Kernel launch
-  if (cuda_kernel_profiler->get_profiler_type() == CUDA_KERNEL_PROFILER_EVENT) {
-    task_handle = profiler->start_with_handle(task_name);
-  } else if (cuda_kernel_profiler->get_profiler_type() ==
-             CUDA_KERNEL_PROFILER_CUPTI) {
-    cuda_kernel_profiler->record_launched_kernel(task_name);
+  if(profiler){
+    profiler->record(task_handle, task_name);
   }
   auto context_guard = CUDAContext::get_instance().get_guard();
 
@@ -103,7 +100,7 @@ void CUDAContext::launch(void *func,
                          shared_mem_bytes, nullptr, arg_pointers.data(),
                          nullptr);
   }
-  if (cuda_kernel_profiler->get_profiler_type() == CUDA_KERNEL_PROFILER_EVENT)
+  if (profiler->get_profiling_tool() == KernelProfilingTool::cuevent)
     profiler->stop(task_handle);
 
   if (get_current_program().config.debug) {
