+++ taichi/inc/kernel_profiling_mode.inc.h
@@ -4,5 +4,3 @@ PER_MODE(disable)        // common      : False
 PER_MODE(enable)         // common      : True  [CUDA:cuEvent]
 PER_MODE(cuda_accurate)  // CUDA        : CUPTI::AccurateMode
 PER_MODE(cuda_detailed)  // CUDA        : CUPTI::DetailedMode
-PER_MODE(extened_mode1)  // ...
-PER_MODE(extened_mode2)  // ...
