+++ taichi/inc/kernel_profiling_tool.inc.h
@@ -1,0 +1,5 @@
+// Supported kernel profiler
+
+PER_TOOL(undef)          // common      : undef
+PER_TOOL(cuevent)        // CUDA        : cuEvent
+PER_TOOL(cupti)          // CUDA        : CUPTI
