+++ taichi/llvm/llvm_program.cpp
@@ -80,13 +80,13 @@ LlvmProgramImpl::LlvmProgramImpl(CompileConfig &config_,
     config_.max_block_dim = 1024;
   }
 
-  if ((config.kernel_profiler != KernelProfilerMode::disable) &&
+  if ((config.kernel_profiler != KernelProfilingMode::disable) &&
       runtime_mem_info) {
     runtime_mem_info->set_profiler(profiler);
   }
   // #if defined(TI_WITH_CUDA)
   //   if (config.arch == Arch::cuda) {
-  //     if (config.kernel_profiler != KernelProfilerMode::disable) {
+  //     if (config.kernel_profiler != KernelProfilingMode::disable) {
   //       CUDAContext::get_instance().set_profiler(profiler);
   //     } else {
   //       CUDAContext::get_instance().set_profiler(nullptr);
