+++ taichi/program/kernel_profiler.cpp
@@ -7,13 +7,13 @@
 
 TLANG_NAMESPACE_BEGIN
 
-std::string kernel_profiler_name(KernelProfilerMode mode) {
+std::string kernel_profiler_name(KernelProfilingMode mode) {
   switch (mode) {
 #define PER_MODE(x)           \
-  case KernelProfilerMode::x: \
+  case KernelProfilingMode::x: \
     return #x;                \
     break;
-#include "taichi/inc/kernel_profiler_mode.inc.h"
+#include "taichi/inc/kernel_profiling_mode.inc.h"
 
 #undef PER_MODE
     default:
@@ -21,15 +21,15 @@ std::string kernel_profiler_name(KernelProfilerMode mode) {
   }
 }
 
-KernelProfilerMode kernel_profiler_from_name(const std::string &mode) {
+KernelProfilingMode kernel_profiler_from_name(const std::string &mode) {
 #define PER_MODE(x)               \
   else if (mode == #x) {          \
-    return KernelProfilerMode::x; \
+    return KernelProfilingMode::x; \
   }
 
   if (false) {
   }
-#include "taichi/inc/kernel_profiler_mode.inc.h"
+#include "taichi/inc/kernel_profiling_mode.inc.h"
 
   else {
     TI_ERROR("Unknown kernel profiler name: {}", mode);
@@ -137,7 +137,7 @@ void KernelProfilerBase::clear() {
   sync();
   total_time_ms = 0;
   records.clear();
-  clear_backend();
+  clear_toolkit();
 }
 
 namespace {
@@ -180,286 +180,10 @@ class DefaultProfiler : public KernelProfilerBase {
   std::string title_;
 };
 
-// A CUDA kernel profiler that uses CUDA timing events
-class KernelProfilerCUDA : public KernelProfilerBase {
- public:
-#if defined(TI_WITH_CUDA)
-  std::map<std::string, std::vector<std::pair<void *, void *>>>
-      outstanding_events;
-#endif
-
-  explicit KernelProfilerCUDA(KernelProfilerMode &mode) {
-#if defined(TI_WITH_CUDA)
-    cuda_profiler = std::make_unique<CUDAProfiler>(mode);
-    CUDAContext::get_instance().set_cuda_profiler(cuda_profiler.get());
-    if (cuda_profiler->get_profiler_type() == CUDA_KERNEL_PROFILER_CUPTI) {
-      cuda_profiler->init_cupti();
-      cuda_profiler->begin_profiling();
-    }
-#endif
-    mode_ = mode;
-  }
-
-  void clear_backend() {
-#if defined(TI_WITH_CUDA)
-    cuda_profiler->clear_traced_records();
-#endif
-  }
-
-  TaskHandle start_with_handle(const std::string &kernel_name) override {
-#if defined(TI_WITH_CUDA)
-    void *start, *stop;
-    CUDADriver::get_instance().event_create(&start, CU_EVENT_DEFAULT);
-    CUDADriver::get_instance().event_create(&stop, CU_EVENT_DEFAULT);
-    CUDADriver::get_instance().event_record(start, 0);
-    outstanding_events[kernel_name].push_back(std::make_pair(start, stop));
-
-    if (!base_event_) {
-      // Note that CUDA driver API only allows querying relative time difference
-      // between two events, therefore we need to manually build a mapping
-      // between GPU and CPU time.
-      // TODO: periodically reinitialize for more accuracy.
-      int n_iters = 100;
-      // Warm up CUDA driver, and use the final event as the base event.
-      for (int i = 0; i < n_iters; i++) {
-        void *e;
-        CUDADriver::get_instance().event_create(&e, CU_EVENT_DEFAULT);
-        CUDADriver::get_instance().event_record(e, 0);
-        CUDADriver::get_instance().event_synchronize(e);
-        auto final_t = Time::get_time();
-        if (i == n_iters - 1) {
-          base_event_ = e;
-          // TODO: figure out a better way to synchronize CPU and GPU time.
-          constexpr float64 cuda_time_offset = 3e-4;
-          // Since event recording and synchronization can take 5 us, it's hard
-          // to exactly measure the real event time. Also note there seems to be
-          // a systematic time offset on CUDA, so adjust for that using a 3e-4 s
-          // magic number.
-          base_time_ = final_t + cuda_time_offset;
-        } else {
-          CUDADriver::get_instance().event_destroy(e);
-        }
-      }
-    }
-
-    return stop;
-#else
-    TI_NOT_IMPLEMENTED;
-#endif
-  }
-
-  virtual void stop(TaskHandle handle) override {
-#if defined(TI_WITH_CUDA)
-    CUDADriver::get_instance().event_record(handle, 0);
-#else
-    TI_NOT_IMPLEMENTED;
-#endif
-  }
-
-  std::string title() const override {
-#if defined(TI_WITH_CUDA)
-    if (cuda_profiler->get_profiling_mode() == KernelProfilerMode::disable)
-      return "Profiler disabled";
-    else if (cuda_profiler->get_profiler_type() == CUDA_KERNEL_PROFILER_EVENT)
-      return "cuEvent Profiler";
-    else if (cuda_profiler->get_profiler_type() == CUDA_KERNEL_PROFILER_CUPTI) {
-      std::string mode = cuda_profiler->get_profiling_mode() ==
-                                 KernelProfilerMode::cuda_accurate
-                             ? "accurate mode"
-                             : "detailed mode";
-      return "nvCUPTI Profiler :: " + mode;
-    }
-#endif
-  }
-
-  void sync() override {
-#if defined(TI_WITH_CUDA)
-    CUDADriver::get_instance().stream_synchronize(nullptr);
-
-    if (cuda_profiler->get_profiler_type() == CUDA_KERNEL_PROFILER_EVENT) {
-      auto &timeline = Timeline::get_this_thread_instance();
-      for (auto &map_elem : outstanding_events) {
-        auto &list = map_elem.second;
-        for (auto &item : list) {
-          auto start = item.first, stop = item.second;
-          float kernel_time;
-          CUDADriver::get_instance().event_elapsed_time(&kernel_time, start,
-                                                        stop);
-
-          if (Timelines::get_instance().get_enabled()) {
-            float time_since_base;
-            CUDADriver::get_instance().event_elapsed_time(&time_since_base,
-                                                          base_event_, start);
-            timeline.insert_event({map_elem.first, true,
-                                   base_time_ + time_since_base * 1e-3,
-                                   "cuda"});
-            timeline.insert_event(
-                {map_elem.first, false,
-                 base_time_ + (time_since_base + kernel_time) * 1e-3, "cuda"});
-          }
-
-          auto it = std::find_if(
-              records.begin(), records.end(),
-              [&](KernelProfileRecord &r) { return r.name == map_elem.first; });
-          if (it == records.end()) {
-            records.emplace_back(map_elem.first);
-            it = std::prev(records.end());
-          }
-          it->insert_sample(kernel_time);
-          total_time_ms += kernel_time;
-
-          // TODO: the following two lines seem to increases profiler overhead a
-          // little bit. Is there a way to avoid the overhead while not creating
-          // too many events?
-          CUDADriver::get_instance().event_destroy(start);
-          CUDADriver::get_instance().event_destroy(stop);
-        }
-      }
-      outstanding_events.clear();
-    } else if (cuda_profiler->get_profiler_type() ==
-               CUDA_KERNEL_PROFILER_CUPTI) {
-      cuda_profiler->calculate_metric_values();
-      cuda_profiler->statistics_on_traced_records(records, total_time_ms);
-      cuda_profiler->end_profiling();
-      cuda_profiler->deinit_cupti();
-      cuda_profiler->init_cupti();
-      cuda_profiler->begin_profiling();
-    }
-#else
-    TI_WARN("Profiler not implemented;");
-#endif
-  }
-
-  void print() override {
-#if defined(TI_WITH_CUDA)
-    sync();
-    if (cuda_profiler->get_profiling_mode() != KernelProfilerMode::disable) {
-      fmt::print("{}\n", title());
-    }
-    if (cuda_profiler->get_profiling_mode() == KernelProfilerMode::enable) {
-      fmt::print(
-          "===================================================================="
-          "===="
-          "=\n");
-      fmt::print(
-          "[      %     total   count |      min       avg       max   ] "
-          "Kernel "
-          "name\n");
-      std::sort(records.begin(), records.end());
-      for (auto &rec : records) {
-        auto fraction = rec.total / total_time_ms * 100.0f;
-        fmt::print(
-            "[{:6.2f}% {:7.3f} s {:6d}x |{:9.3f} {:9.3f} {:9.3f} ms] {}\n",
-            fraction, rec.total / 1000.0f, rec.counter, rec.min,
-            rec.total / rec.counter, rec.max, rec.name);
-      }
-      fmt::print(
-          "--------------------------------------------------------------------"
-          "----"
-          "-\n");
-      fmt::print(
-          "[100.00%] Total kernel execution time: {:7.3f} s   number of "
-          "records: "
-          "{}\n",
-          get_total_time(), records.size());
-
-      fmt::print(
-          "===================================================================="
-          "===="
-          "=\n");
-    } else if (cuda_profiler->get_profiling_mode() ==
-               KernelProfilerMode::cuda_accurate) {
-      fmt::print(
-          "===================================================================="
-          "=============================="
-          "=\n");
-      fmt::print(
-          "[      %     total   count |      min       avg       max    |    "
-          "load.g   store.g    ] Kernel "
-          "name\n");
-      std::sort(records.begin(), records.end());
-      for (auto &rec : records) {
-        auto fraction = rec.total / total_time_ms * 100.0f;
-        fmt::print(
-            "[{:6.2f}% {:7.3f} s {:6d}x |{:9.3f} {:9.3f} {:9.3f} ms | {:9.3f} "
-            "{:9.3f}  MB] {}\n",
-            fraction, rec.total / 1000.0f, rec.counter, rec.min,
-            rec.total / rec.counter, rec.max,
-            rec.mem_load_in_bytes / rec.counter / 1024 / 1024,
-            rec.mem_store_in_bytes / rec.counter / 1024 / 1024, rec.name);
-      }
-      fmt::print(
-          "--------------------------------------------------------------------"
-          "------------------------------"
-          "-\n");
-      fmt::print(
-          "[100.00%] Total kernel execution time: {:7.3f} s   number of "
-          "records: "
-          "{}\n",
-          get_total_time(), records.size());
-
-      fmt::print(
-          "===================================================================="
-          "=============================="
-          "=\n");
-    } else if (cuda_profiler->get_profiling_mode() ==
-               KernelProfilerMode::cuda_detailed) {
-      fmt::print(
-          "===================================================================="
-          "========================================================"
-          "=\n");
-      fmt::print(
-          "[      %     total   count |      min       avg       max    |    "
-          "load.g   store.g     |  uti.core   uti.dram ] Kernel "
-          "name\n");
-      std::sort(records.begin(), records.end());
-      for (auto &rec : records) {
-        auto fraction = rec.total / total_time_ms * 100.0f;
-        fmt::print(
-            "[{:6.2f}% {:7.3f} s {:6d}x |{:9.3f} {:9.3f} {:9.3f} ms | {:9.3f} "
-            "{:9.3f}  MB |    {:2.2f}%     {:2.2f}% ] {}\n",
-            fraction, rec.total / 1000.0f, rec.counter, rec.min,
-            rec.total / rec.counter, rec.max,
-            rec.mem_load_in_bytes / rec.counter / 1024 / 1024,
-            rec.mem_store_in_bytes / rec.counter / 1024 / 1024,
-            rec.utilization_core / rec.counter,
-            rec.utilization_mem / rec.counter, rec.name);
-      }
-      fmt::print(
-          "--------------------------------------------------------------------"
-          "--------------------------------------------------------"
-          "-\n");
-      fmt::print(
-          "[100.00%] Total kernel execution time: {:7.3f} s   number of "
-          "records: "
-          "{}\n",
-          get_total_time(), records.size());
-
-      fmt::print(
-          "===================================================================="
-          "========================================================"
-          "=\n");
-    }
-#endif
-  }
-
-  // deprecated
-  // static KernelProfilerCUDA &get_instance() {
-  //   static KernelProfilerCUDA profiler;
-  //   return profiler;
-  // }
-
- private:
-  void *base_event_{nullptr};
-  float64 base_time_{0.0};
-#if defined(TI_WITH_CUDA)
-  std::unique_ptr<CUDAProfiler> cuda_profiler{nullptr};
-#endif
-};
 }  // namespace
 
 std::unique_ptr<KernelProfilerBase> make_profiler(Arch arch,
-                                                  KernelProfilerMode &mode) {
+                                                  KernelProfilingMode &mode) {
   if (arch == Arch::cuda) {
     return std::make_unique<KernelProfilerCUDA>(mode);
   } else {
