+++ taichi/python/export_lang.cpp
@@ -77,9 +77,9 @@ void export_lang(py::module &m) {
   m.def("arch_name", arch_name);
   m.def("arch_from_name", arch_from_name);
 
-  py::enum_<KernelProfilerMode>(m, "KernelProfilerMode", py::arithmetic())
-#define PER_MODE(x) .value(#x, KernelProfilerMode::x)
-#include "taichi/inc/kernel_profiler_mode.inc.h"
+  py::enum_<KernelProfilingMode>(m, "KernelProfilingMode", py::arithmetic())
+#define PER_MODE(x) .value(#x, KernelProfilingMode::x)
+#include "taichi/inc/kernel_profiling_mode.inc.h"
 #undef PER_MODE
       .export_values();
